chore: add ci badge
